feat: lazy load variables page
