+++ src/components/variable-grid.tsx
